requirements.txt
Flask==2.1.1
Werkzeug==2.1.1
pydub==0.25.1
Pillow==9.1.0
moviepy==1.0.3
pdf2image==1.16.0
reportlab==3.6.6
apscheduler==3.9.1
